removed aptos dependencies, updated example apps, cleaned exports
