+++ packages/supra-starkey-connect/lib/global.d.ts
@@ -1,5 +1,5 @@
 import type { StarkeyObject } from './types'
-import type { SupraStarkeyConnect } from './ssc'
+import type { SupraStarkeyConnect } from './lib'
 declare global {
   interface Window {
     starkey?: StarkeyObject
